refactor(user): improve error handling and validation messages in UserController and StoreUserRequest

+++ app/Http/Controllers/UserController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Http\Requests\{PaginationRequest, StoreUserRequest, UpdateUserRequest};
 use App\Models\User;
 use App\Services\UserService;
@@ -19,45 +18,44 @@ class UserController extends Controller
 
     public function store(StoreUserRequest $request)
     {
-        try {
-            $user = $this->service->create($request->validated());
+        $user = $this->service->create($request->validated());
 
-            return response()->json(['data' => $user], 201);
-        } catch (\Exception $e) {
+        if (!$user) {
             return response()->json(['error' => 'User creation failed'], 500);
         }
+
+        return response()->json(['data' => $user], 201);
     }
 
     public function show(User $user)
     {
-        try {
-            return response()->json(['data' => $user], 200);
-        } catch (\Exception $e) {
+        if (!$user) {
             return response()->json(['error' => 'User not found'], 404);
         }
+        return response()->json(['data' => $user], 200);
     }
 
     public function update(UpdateUserRequest $request, User $user)
     {
-        try {
-            return response()->json(
-                $this->service->update($user, $request->validated()),
-                200
-            );
-        } catch (\Exception $e) {
-            return response()->json(['error' => 'User update failed'], 400);
+
+        if (!$user) {
+            return response()->json(['error' => 'User not found'], 404);
         }
+        return response()->json(
+            $this->service->update($user, $request->validated()),
+            200
+        );
     }
+
 
     public function destroy(User $user)
     {
-        try {
 
-            $user->delete();
-
-            return response()->json(null, 204);
-        } catch (\Exception $e) {
-            return response()->json(['error' => 'User deletion failed'], 400);
+        if (!$user) {
+            return response()->json(['error' => 'User not found'], 404);
         }
+        $user->delete();
+
+        return response()->json(null, 204);
     }
 }
